LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-expose-star.sub.any-expected.txt


FAIL Basic Access-Control-Expose-Headers: * support assert_equals: expected (string) "X" but got (object) null
PASS * for credentialed fetches only matches literally 
FAIL * can be one of several values assert_equals: expected (string) "X" but got (object) null
